% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{adjustGDPComponent}
\alias{adjustGDPComponent}
\title{Adjust EmpComp, Tax, and GOS to fill NA and make them consistent with GDP}
\usage{
adjustGDPComponent(year, return)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{return}{A character string showing which attribute to return. 'comp', 'tax', 'gos'}
}
\value{
A data frame contains adjusted EmpComp, Tax, GOS, and GDP
}
\description{
Adjust EmpComp, Tax, and GOS to fill NA and make them consistent with GDP
}
