% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{adjustUseTablebyImportMatrix}
\alias{adjustUseTablebyImportMatrix}
\title{Adjust US Use table by Import matrix.}
\usage{
adjustUseTablebyImportMatrix(iolevel, year)
}
\arguments{
\item{iolevel}{Level of detail, can be "Sector", "Summary, "Detail".}

\item{year}{A numeric value specifying the year of interest.}
}
\value{
A adjusted US Use table.
}
\description{
Adjust US Use table by Import matrix.
}
