% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{allocateStateTabletoBEASummary}
\alias{allocateStateTabletoBEASummary}
\title{Allocate state table (GDP, Tax, Employment Compensation, and GOS) to BEA Summary based on specified weight}
\usage{
allocateStateTabletoBEASummary(statetablename, year, allocationweightsource)
}
\arguments{
\item{statetablename}{Name of pre-saved state table, canbe GDP, Tax, Employment Compensation, and GOS.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{allocationweightsource}{Source of allocatino weight, can be "Employment".}
}
\value{
A data frame contains allocated state value for all states with row names being BEA sector code.
}
\description{
Allocate state table (GDP, Tax, Employment Compensation, and GOS) to BEA Summary based on specified weight
}
