% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{AlternativeStateCommodityOutputRatio_2012}
\alias{AlternativeStateCommodityOutputRatio_2012}
\title{State commodity output ratios from alternative sources 2012}
\format{
A dataframe with 1198 obs. and 4 variables
\describe{
 \item{BEA_2012_Summary_Code}{character}
 \item{State}{character}
 \item{Value}{numeric}
 \item{Ratio}{numeric}
}
}
\usage{
AlternativeStateCommodityOutputRatio_2012
}
\description{
State commodity output ratios from alternative sources 2012
}
\keyword{datasets}
