% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{AlternativeStateCommodityOutputRatio_2017}
\alias{AlternativeStateCommodityOutputRatio_2017}
\title{State commodity output ratios from alternative sources 2017}
\format{
A dataframe with 1208 obs. and 4 variables
\describe{
 \item{BEA_2012_Summary_Code}{character}
 \item{State}{character}
 \item{Value}{numeric}
 \item{Ratio}{numeric}
}
}
\usage{
AlternativeStateCommodityOutputRatio_2017
}
\description{
State commodity output ratios from alternative sources 2017
}
\keyword{datasets}
