% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{assembleStateValueAdded}
\alias{assembleStateValueAdded}
\title{Assemble Summary-level value added sectors (V001, V002, V003) for all states at a specific year.}
\usage{
assembleStateValueAdded(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains Summary-level value added (V001, V002, V003) for all states at a specific year.
}
\description{
Assemble Summary-level value added sectors (V001, V002, V003) for all states at a specific year.
}
