% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{assembleTwoRegionIO}
\alias{assembleTwoRegionIO}
\title{Assemble two-region make, use, domestic use, and demand tables as well as industry and commodity outputs.}
\usage{
assembleTwoRegionIO(iolevel, year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A list of two-region make, use, domestic use, and demand tables
as well as industry and commodity outputs by state.
}
\description{
Assemble two-region make and use tables as well as industry and commodity outputs.
}
