% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildFullTwoRegionIOTable}
\alias{buildFullTwoRegionIOTable}
\title{Build a full two-region IO table for specified state and rest of US for a given year.}
\usage{
buildFullTwoRegionIOTable(state, year, iolevel, specs)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, currently can only be "Summary",
theoretically can be "Detail", or "Sector" in future versions.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A full two-region IO table for specified state and rest of US for a given year.
}
\description{
Build a full two-region IO table for specified state and rest of US for a given year.
}
