% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildFullTwoRegionStateModel}
\alias{buildFullTwoRegionStateModel}
\title{Build a two-region state model for all 52 states/regions (including DC and Overseas) for a given year}
\usage{
buildFullTwoRegionStateModel(state, year, ioschema, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{ioschema}{A numeric value of either 2012 or 2007 specifying the io schema year.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A list of state demand model components: Use table and Domestic Use table.
}
\description{
Build a state demand model for all 52 states/regions (including DC and Overseas) for a given year.
}
