% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildStateDemandModel}
\alias{buildStateDemandModel}
\title{Build a state demand model for all 52 states/regions
(including DC and Overseas) for a given year}
\usage{
buildStateDemandModel(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A list of state demand model components: Use table and Domestic Use table.
}
\description{
Build a state demand model for all 52 states/regions
(including DC and Overseas) for a given year.
}
