% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildTwoRegionDemandModel}
\alias{buildTwoRegionDemandModel}
\title{Build a two-region demand model}
\usage{
buildTwoRegionDemandModel(
  state,
  year,
  ioschema,
  iolevel,
  ICF_sensitivity_analysis = FALSE,
  adjust_by = 0,
  domestic = TRUE
)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{ioschema}{A numeric value of either 2012 or 2007 specifying the io schema year.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A list of domestic two-region demand tables.
}
\description{
Generate two-region (SoI and RoUS) demand tables
with interregional exports and imports.
}
