% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommodityFlowFunctions.R
\name{calculateCommodityFlowRatios}
\alias{calculateCommodityFlowRatios}
\title{Calculate domestic/import/export commodity flow ratios by state}
\usage{
calculateCommodityFlowRatios(state, year, flow_ratio_type, specs, iolevel)
}
\arguments{
\item{state}{State name.}

\item{year}{A numeric value specifying the year of interest.}

\item{flow_ratio_type}{Type of commodity flow, can be "domestic", "export", or "import".}

\item{specs}{A list of model specs including 'BaseIOSchema'}

\item{iolevel}{BEA sector level of detail, currently can only be "Summary",
theoretically can be "Detail", or "Sector" in future versions.}
}
\value{
A data frame contains commodity flow ratios by BEA.
}
\description{
Calculate domestic/import/export commodity flow ratios by state
}
