% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommodityFlowFunctions.R
\name{calculateHazWasteManagementServiceFlowRatios}
\alias{calculateHazWasteManagementServiceFlowRatios}
\title{Calculate domestic hazardous waste management services flow ratios by state}
\usage{
calculateHazWasteManagementServiceFlowRatios(state, year, specs)
}
\arguments{
\item{state}{State name.}

\item{year}{A numeric value between 2012 and 2017 specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains hazardous waste management services flow ratios by BEA.
}
\description{
Calculate domestic hazardous waste management services flow ratios by state
}
