% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{calculateRowColumnDiffernce}
\alias{calculateRowColumnDiffernce}
\title{calculateRowColumnDiffernce (MODIFIED)}
\usage{
calculateRowColumnDiffernce(matrix, t_cs, t_rs)
}
\arguments{
\item{matrix}{Matrix, matrix to be processed}

\item{t_cs}{Vector, true column sum}

\item{t_rs}{Vector, true row sum}
}
\value{
a list containing row difference and column difference
}
\description{
This function is to return a list containing row difference and column difference of a 
matrix with NA
}
