% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{calculateStateCommodityOutputRatio}
\alias{calculateStateCommodityOutputRatio}
\title{Calculate state-US Commodity Output ratios at BEA Summary level.}
\usage{
calculateStateCommodityOutputRatio(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state-US Commodity Output ratios at BEA Summary level.
}
\description{
Calculate state-US Commodity Output ratios at BEA Summary level.
}
