% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{calculateStateIndustryCommodityOuput}
\alias{calculateStateIndustryCommodityOuput}
\title{Calculate output from model objects}
\usage{
calculateStateIndustryCommodityOuput(model)
}
\arguments{
\item{model}{An stateior model object with model specs and specific IO tables loaded}
}
\value{
A model object with disaggregated IndustryOutput and CommodityOutput objects
}
\description{
Calculate output from model objects
}
