% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{calculateStateIndustryOutputbyLineCode}
\alias{calculateStateIndustryOutputbyLineCode}
\title{Calculate state industry output by BEA State LineCode
by multiplying state_US_VA_ratio_LineCode by USGrossOutput_LineCode.}
\usage{
calculateStateIndustryOutputbyLineCode(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains state industry output by BEA State LineCode.
}
\description{
Calculate state industry output by BEA State LineCode
by multiplying state_US_VA_ratio_LineCode by USGrossOutput_LineCode.
}
