% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{calculateStateSLGovExpenditureRatio}
\alias{calculateStateSLGovExpenditureRatio}
\title{Calculate state S&L government expenditure ratio at BEA Summary level.}
\usage{
calculateStateSLGovExpenditureRatio(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'model_ver'}
}
\value{
A data frame contains state S&L government expenditure ratio for all states at a specific year.
}
\description{
Calculate state S&L government expenditure ratio at BEA Summary level.
}
