% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{calculateStatetoBEASummaryAllocationFactor}
\alias{calculateStatetoBEASummaryAllocationFactor}
\title{Calculate allocation factors based on state-level data, such as employment}
\usage{
calculateStatetoBEASummaryAllocationFactor(year, allocationweightsource, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{allocationweightsource}{A string specifying the source being used
as the weight in the allocation.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains allocation factors
for all states with row names being BEA sector code.
}
\description{
Calculate allocation factors based on state-level data, such as employment
}
