% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{calculateStateTotalPCE}
\alias{calculateStateTotalPCE}
\title{Calculate state total PCE (personal consumption expenditures) at BEA Summary level.}
\usage{
calculateStateTotalPCE(year)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}
}
\value{
A data frame contains ratios of statetotal PCE for all states at a specific year at BEA Summary level.
}
\description{
Calculate state total PCE (personal consumption expenditures) at BEA Summary level.
}
