% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{calculateStateUSEmpCompensationRatio}
\alias{calculateStateUSEmpCompensationRatio}
\title{Calculate state-US employee compensation ratios at BEA Summary level.}
\usage{
calculateStateUSEmpCompensationRatio(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains state-US employment compensation ratios
for all states at a specific year at BEA Summary level.
}
\description{
Calculate state-US employee compensation ratios at BEA Summary level.
}
