% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{calculateStateUSPCERatio}
\alias{calculateStateUSPCERatio}
\title{Calculate state-US PCE (personal consumption expenditures) ratios at BEA Summary level.}
\usage{
calculateStateUSPCERatio(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains ratios of state/US PCE for all states at a specific year at BEA Summary level.
}
\description{
Calculate state-US PCE (personal consumption expenditures) ratios at BEA Summary level.
}
