% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{calculateStateUSVARatiobyLineCode}
\alias{calculateStateUSVARatiobyLineCode}
\title{Calculate state-US GVA (value added) ratios by BEA State LineCode.}
\usage{
calculateStateUSVARatiobyLineCode(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains ratios of state/US GVA (value added)
for all states at a specific year by BEA State LineCode.
}
\description{
Calculate state-US GVA (value added) ratios by BEA State LineCode.
}
