% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{calculateUSDomesticUseRatioMatrix}
\alias{calculateUSDomesticUseRatioMatrix}
\title{Calculate US Domestic Use Ratio (matrix).}
\usage{
calculateUSDomesticUseRatioMatrix(iolevel, year)
}
\arguments{
\item{iolevel}{Level of detail, can be "Sector", "Summary, "Detail".}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains US Domestic Use Ratio (matrix) at a specific year at BEA Summary level.
}
\description{
Calculate US Domestic Use Ratio (matrix).
}
