% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{calculateUSGovExpenditureWeightFactor}
\alias{calculateUSGovExpenditureWeightFactor}
\title{Calculate weighting factor of each expenditure component over US total gov expenditure.}
\usage{
calculateUSGovExpenditureWeightFactor(year, defense)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{defense}{A logical value indicating if the expenditure is spent on defense or not.}
}
\value{
A data frame contains weighting factor of each expenditure component over US total gov expenditure.
}
\description{
Calculate weighting factor of each expenditure component over US total gov expenditure.
}
