% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{calculateUSImportRatioMatrix}
\alias{calculateUSImportRatioMatrix}
\title{Calculate US Import Ratio (matrix) at BEA Summary level.}
\usage{
calculateUSImportRatioMatrix(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains US Import Ratio (matrix) at a specific year at BEA Summary level.
}
\description{
Calculate US Import Ratio (matrix) at BEA Summary level.
}
