% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{calculateUSInternationalTransportMarginsRatioMatrix}
\alias{calculateUSInternationalTransportMarginsRatioMatrix}
\title{Calculate US International Transport Margins Ratio (matrix).}
\usage{
calculateUSInternationalTransportMarginsRatioMatrix(iolevel, year, specs)
}
\arguments{
\item{iolevel}{Level of detail, can be "Sector", "Summary, "Detail".}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains US International Transport Margins Ratio (matrix) at a specific year at BEA Summary level.
}
\description{
Calculate US International Transport Margins Ratio (matrix).
}
