% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommodityFlowFunctions.R
\name{calculateUtilitiesFlowRatios}
\alias{calculateUtilitiesFlowRatios}
\title{Calculate domestic interregional utilities flow ratios by state,
utilities include electricity generation, transmission and distribution,
natural gas distribution and water, sewage and other}
\usage{
calculateUtilitiesFlowRatios(state, year)
}
\arguments{
\item{state}{State name.}

\item{year}{A numeric value between 2012 and 2017 specifying the year of interest.}
}
\value{
A data frame contains domestic interregional utilities flow ratios by state.
}
\description{
Calculate domestic interregional utilities flow ratios by state,
utilities include electricity generation, transmission and distribution,
natural gas distribution and water, sewage and other
}
