% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Census_StateExport_2013}
\alias{Census_StateExport_2013}
\title{State Export table (2013)}
\format{
A data frame with 8554 obs. and 5 variables
\describe{
 \item{NAICS}{text code}
 \item{CTY_NAME}{text code}
 \item{STATE}{text code}
 \item{ALL_VAL_YR}{numeric value}
 \item{YEAR}{integer}
}
}
\source{
\url{https://api.census.gov/data/timeseries/intltrade/exports}
}
\usage{
Census_StateExport_2013
}
\description{
State Export table (2013)
}
\keyword{datasets}
