% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Census_StateExport_2018}
\alias{Census_StateExport_2018}
\title{State Export table (2018)}
\format{
A data frame with 7923 obs. and 5 variables
\describe{
 \item{NAICS}{text code}
 \item{CTY_NAME}{text code}
 \item{STATE}{text code}
 \item{ALL_VAL_YR}{numeric value}
 \item{YEAR}{integer}
}
}
\source{
\url{https://api.census.gov/data/timeseries/intltrade/exports}
}
\usage{
Census_StateExport_2018
}
\description{
State Export table (2018)
}
\keyword{datasets}
