% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Census_StateImport_2016}
\alias{Census_StateImport_2016}
\title{State Import table (2016)}
\format{
A data frame with 8009 obs. and 5 variables
\describe{
 \item{NAICS}{text code}
 \item{CTY_NAME}{text code}
 \item{STATE}{text code}
 \item{GEN_VAL_YR}{numeric value}
 \item{YEAR}{integer}
}
}
\source{
\url{https://api.census.gov/data/timeseries/intltrade/imports}
}
\usage{
Census_StateImport_2016
}
\description{
State Import table (2016)
}
\keyword{datasets}
