% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Census_StateLocalGovExpenditure_2013}
\alias{Census_StateLocalGovExpenditure_2013}
\title{State and Local Government Expenditure table (2013)}
\format{
A data frame with 86 obs. and 53 variables (including 50 states and D.C.)
\describe{
 \item{Description}{text code}
 \item{United States Total}{current year USD}
 \item{Alabama}{current year USD}
 ...
 \item{Wyoming}{current year USD}
}
}
\source{
\url{https://www.census.gov/programs-surveys/gov-finances/data/datasets.All.html}
}
\usage{
Census_StateLocalGovExpenditure_2013
}
\description{
State and Local Government Expenditure table (2013)
}
\keyword{datasets}
