% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Census_USATradeImport_2012}
\alias{Census_USATradeImport_2012}
\title{State Import table (2012)}
\format{
A data frame with 6834 obs. and 4 variables
\describe{
 \item{State}{text code}
 \item{Year}{integer}
 \item{NAICS}{integer}
 \item{Value}{numeric value}
}
}
\source{
\url{https://usatrade.census.gov}
}
\usage{
Census_USATradeImport_2012
}
\description{
State Import table (2012)
}
\keyword{datasets}
