% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{checkFileonDataCommons}
\alias{checkFileonDataCommons}
\title{Check if file is available on Data Commons. Stop function execution if not.}
\usage{
checkFileonDataCommons(file)
}
\arguments{
\item{filename}{A string specifying filename "State_Summary_Use_2017".}
}
\description{
Check if file is available on Data Commons. Stop function execution if not.
}
