% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{checkIOLevel}
\alias{checkIOLevel}
\title{Check IO level. Stop function execution if input parameter is not "Summary".}
\usage{
checkIOLevel(iolevel)
}
\arguments{
\item{iolevel}{BEA sector level of detail, currently can only be "Summary",
theoretically can be "Detail", or "Sector" in future versions.}
}
\description{
Check IO level. Stop function execution if input parameter is not "Summary".
}
