% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{checkStateName}
\alias{checkStateName}
\title{Check state name. Stop function execution if input parameter is not a valid state name.}
\usage{
checkStateName(state)
}
\arguments{
\item{state}{A text value specifying state of interest.}
}
\description{
Check state name. Stop function execution if input parameter is not a valid state name.
}
