% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{createDisaggFilesFromProxyData}
\alias{createDisaggFilesFromProxyData}
\title{Create Make, Use, and Env ratio files for each state from Proxy data for the relevant sectors.}
\usage{
createDisaggFilesFromProxyData(model, disagg, disaggYear, disaggState)
}
\arguments{
\item{model}{An stateior model object with model specs and specific IO tables loaded}

\item{disagg}{Specifications for disaggregating the current Table}

\item{disaggYear}{Integer specifying the state model year}

\item{disaggState}{A string value that indicates the state model being disaggregated. For national models, string should be "US"}
}
\value{
A stateior model with disaggregation specs included
}
\description{
Create Make, Use, and Env ratio files for each state from Proxy data for the relevant sectors.
}
