% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{createMatrixForRASM0}
\alias{createMatrixForRASM0}
\title{createMatrixForRASM0 (MODIFIED)}
\usage{
createMatrixForRASM0(matrixKEY, matrix)
}
\arguments{
\item{matrixKEY}{Matrix, a boolean matrix = is.na(matrix)}

\item{matrix}{Matrix, matrix to be processed}
}
\value{
a matrix M0 for RAS
}
\description{
This function is to create a matrix that with 0 value at positions with all known value
and retains estimated value at positions with NA (M0 for RAS method)
}
