% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{disaggregateNationalObjectsInStateModel}
\alias{disaggregateNationalObjectsInStateModel}
\title{Disaggregate national make and use tables}
\usage{
disaggregateNationalObjectsInStateModel(model, disagg)
}
\arguments{
\item{model}{An stateior model object with model specs and specific IO tables loaded}

\item{disagg}{Specifications for model disaggregation}
}
\value{
A stateior model with the disaggregateed objects
}
\description{
Disaggregate national make and use tables
}
