% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{disaggregateStateModel}
\alias{disaggregateStateModel}
\title{Disaggregate state make and use tables}
\usage{
disaggregateStateModel(model, state)
}
\arguments{
\item{model}{An stateior model object with model specs and specific IO tables loaded}

\item{state}{A string value that indicates the state model being disaggregated}
}
\value{
A stateior model with the disaggregateed objects
}
\description{
Disaggregate state make and use tables
}
