% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{disaggregateStateSectorLists}
\alias{disaggregateStateSectorLists}
\title{Disaggregate model$Commodity or model$Industry dataframes in the main model object}
\usage{
disaggregateStateSectorLists(code_vector, disagg)
}
\arguments{
\item{code_vector}{A list of sector codes (industry or commodity)}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
newList A list which contain the disaggregated model$Commodity or model$Industry objects
}
\description{
Disaggregate model$Commodity or model$Industry dataframes in the main model object
}
