% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{downloadStateIODatafromDataCommons}
\alias{downloadStateIODatafromDataCommons}
\title{Download StateIO data file from Data Commons and stores in a local data directory.}
\usage{
downloadStateIODatafromDataCommons(filename, ver = NULL)
}
\arguments{
\item{filename}{A string specifying filename, e.g. "State_Summary_Use_2017".}

\item{ver}{A string specifying version of the data, default is NULL, can be "v0.1.0".}
}
\value{
An .rds data file downloaded from Data Commons and stored in local directory.
}
\description{
Download StateIO data file from Data Commons and stores in a local data directory.
}
