% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{estimateStateCommodityOutputRatiofromAlternativeSources}
\alias{estimateStateCommodityOutputRatiofromAlternativeSources}
\title{Estimate state commodity output and ratios from alternative sources, including
USDA Census of Agriculture, NOAA Fisheries, USFS Forestry Inventory, and ORNL
Feight Analysis Framework (FAF).}
\usage{
estimateStateCommodityOutputRatiofromAlternativeSources(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema',}
}
\value{
A data frame contains state commodity output from alternative sources
and calculated state/US commodity ratios for each state.
}
\description{
Estimate state commodity output and ratios from alternative sources, including
USDA Census of Agriculture, NOAA Fisheries, USFS Forestry Inventory, and ORNL
Feight Analysis Framework (FAF).
}
