% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{estimateStateExport}
\alias{estimateStateExport}
\title{Estimate state export at BEA Summary level.}
\usage{
estimateStateExport(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state export for all states at a specific year at BEA Summary level.
}
\description{
Estimate state export at BEA Summary level.
}
