% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{estimateStateIntermediateConsumption}
\alias{estimateStateIntermediateConsumption}
\title{Estimate state Intermediate Consumption at BEA Summary level.
For each state and industry, calculate state_US_IndustryOutput_ratio,
then multiply US_Use_Intermediate by the ratio to get State_Use_Intermediate.}
\usage{
estimateStateIntermediateConsumption(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A list of data.frames containing state Intermediate Consumption.
}
\description{
Estimate state Intermediate Consumption at BEA Summary level.
For each state and industry, calculate state_US_IndustryOutput_ratio,
then multiply US_Use_Intermediate by the ratio to get State_Use_Intermediate.
}
