% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{estimateStateSLGovExpenditure}
\alias{estimateStateSLGovExpenditure}
\title{Estimate state S&L government expenditure at BEA Summary level.}
\usage{
estimateStateSLGovExpenditure(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains state S&L government expenditure for all states
at a specific year at BEA Summary level.
}
\description{
Estimate state S&L government expenditure at BEA Summary level.
}
