% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{estimateUSDomesticUse}
\alias{estimateUSDomesticUse}
\title{Estimate US domestic Use table by adjusting US Use table based on Import matrix.}
\usage{
estimateUSDomesticUse(iolevel, year)
}
\arguments{
\item{iolevel}{Level of detail, can be "Sector", "Summary, "Detail".}

\item{year}{A numeric value specifying the year of interest.}
}
\value{
A US Domestic Use table.
}
\description{
Estimate US domestic Use table by adjusting US Use table based on Import matrix.
}
