% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{FAF_2017}
\alias{FAF_2017}
\title{Oak Ridge National Laboratory (ORNL) FAF table for 2017
A dataset of domestic and international commodity movement
by all modes of transportation.
Data dictionary can be found in Table 5
at https://faf.ornl.gov/fafweb/data/FAF4\%20User\%20Guide.pdf.}
\format{
A data frame with 560968 obs. and 12 variables:
\describe{
 \item{fr_orig}{integer}
 \item{dms_origst}{integer}
 \item{dms_destst}{integer}
 \item{fr_dest}{integer}
 \item{fr_inmode}{integer}
 \item{dms_mode}{integer}
 \item{fr_outmode}{integer}
 \item{sctg2}{integer}
 \item{trade_type}{integer}
 \item{tons_2017}{numeric}
 \item{value_2017}{numeric}
 \item{tmiles_2017}{numeric}
}
}
\source{
\url{https://faf.ornl.gov/fafweb/}
}
\usage{
FAF_2017
}
\description{
Oak Ridge National Laboratory (ORNL) FAF table for 2017
A dataset of domestic and international commodity movement
by all modes of transportation.
Data dictionary can be found in Table 5
at https://faf.ornl.gov/fafweb/data/FAF4\%20User\%20Guide.pdf.
}
\keyword{datasets}
