% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{FedGovExp_EquipmentNonDefense_2012}
\alias{FedGovExp_EquipmentNonDefense_2012}
\title{Federal government expenditure on non-defense equipment final demand in 2012.
Categorized by 6-digit NAICS, by state (including District of Columbia) and
by zipcode.}
\format{
A data frame with 820 obs. and 5 variables:
\describe{
 \item{NAICS}{integer}
 \item{Year}{integer}
 \item{State}{text}
 \item{Zipcode}{integer}
 \item{Amount}{numeric}
}
}
\source{
\url{https://fedspendingtransparency.github.io/}
}
\usage{
FedGovExp_EquipmentNonDefense_2012
}
\description{
Federal government expenditure on non-defense equipment final demand in 2012.
Categorized by 6-digit NAICS, by state (including District of Columbia) and
by zipcode.
}
\keyword{datasets}
