% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{FedGovExp_StructureNonDefense_2012}
\alias{FedGovExp_StructureNonDefense_2012}
\title{Federal government expenditure on non-defense structure final demand in 2012.
Categorized by 6-digit NAICS, by state (including District of Columbia) and
by county.}
\format{
A data frame with 8216 obs. and 5 variables:
\describe{
 \item{Year}{integer}
 \item{NAICS}{integer}
 \item{State}{text}
 \item{County}{text}
 \item{Amount}{numeric}
}
}
\source{
\url{https://www.usaspending.gov/#/download_center/award_data_archive}
}
\usage{
FedGovExp_StructureNonDefense_2012
}
\description{
Federal government expenditure on non-defense structure final demand in 2012.
Categorized by 6-digit NAICS, by state (including District of Columbia) and
by county.
}
\keyword{datasets}
