% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{fillNAwithRatioMatrix}
\alias{fillNAwithRatioMatrix}
\title{fillNAwithRatioMatrix (MODIFIED)}
\usage{
fillNAwithRatioMatrix(matrix_to_fill, ratio_matrix, row_difference)
}
\arguments{
\item{matrix_to_fill}{Matrix, matrix to be processed}

\item{ratio_matrix}{Matrix, ratio matrix}

\item{row_difference}{Vector, row difference}
}
\value{
a matrix whose row sums equal trus row sum
}
\description{
This function is to fill a matrix with NAs by another ratio matrix to neutralize
row difference
}
