% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{findLatestStateIODatainLocalDirectory}
\alias{findLatestStateIODatainLocalDirectory}
\title{Find the latest StateIO data in local data directory.}
\usage{
findLatestStateIODatainLocalDirectory(filename)
}
\arguments{
\item{filename}{A string specifying filename, e.g. "State_Summary_Use_2017".}
}
\value{
File name of the latest StateIO data in local data directory.
}
\description{
Find the latest StateIO data in local data directory.
}
