% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{findLatestStateIODataonDataCommons}
\alias{findLatestStateIODataonDataCommons}
\title{Find the latest state IO data on Data Commons.}
\usage{
findLatestStateIODataonDataCommons(filename)
}
\arguments{
\item{filename}{A string specifying filename "State_Summary_Use_2017".}
}
\value{
File name of the latest state IO data on Data Commons.
}
\description{
Find the latest state IO data on Data Commons.
}
