% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{formatFullUseFromStateToUSEEIO}
\alias{formatFullUseFromStateToUSEEIO}
\title{Format Use table objects to prepare for use in useeior functions}
\usage{
formatFullUseFromStateToUSEEIO(table)
}
\arguments{
\item{table}{FullUse table}
}
\value{
The FullUse model object with useeior formatting fit for disaggregation functions
}
\description{
Format Use table objects to prepare for use in useeior functions
}
