% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{formatFullUseFromUSEEIOtoState}
\alias{formatFullUseFromUSEEIOtoState}
\title{Format Use table objects back to stateior format}
\usage{
formatFullUseFromUSEEIOtoState(model, state, domestic = FALSE)
}
\arguments{
\item{model}{An stateior model object with model specs and specific IO tables loaded}

\item{state}{A string value that indicates the state model being disaggregated}

\item{domestic}{A boolean that indicates whether the table to format is the domesticUse table or not}
}
\value{
A stateior FullUse table formatted according to stateior specifications
}
\description{
Format Use table objects back to stateior format
}
