% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{formatMakeFromStateToUSEEIO}
\alias{formatMakeFromStateToUSEEIO}
\title{Format model objects to prepare for use in useeior functions}
\usage{
formatMakeFromStateToUSEEIO(model, state)
}
\arguments{
\item{model}{An stateior model object with model specs and specific IO tables loaded}

\item{state}{A string value that indicates the state model being disaggregated}
}
\value{
A stateior make table formatted for disaggregation with useeior functions
}
\description{
Format model objects to prepare for use in useeior functions
}
