% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{formatMakeFromUSEEIOtoState}
\alias{formatMakeFromUSEEIOtoState}
\title{Format Make table objects back to stateior format}
\usage{
formatMakeFromUSEEIOtoState(model, state)
}
\arguments{
\item{model}{An stateior model object with model specs and specific IO tables loaded}

\item{state}{A string value that indicates the state model being disaggregated}
}
\value{
A stateior make table formatted according to stateior specifications
}
\description{
Format Make table objects back to stateior format
}
