% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteregionalCommodityFlowFunctions.R
\name{generateDomestic2RegionICFs}
\alias{generateDomestic2RegionICFs}
\title{Generate domestic 2 region inter-regional commodity flows (ICFs) table.}
\usage{
generateDomestic2RegionICFs(state, year, ioschema, iolevel)
}
\arguments{
\item{state}{State name.}

\item{year}{A numeric value between 2012 and 2017 specifying the year of interest.}

\item{ioschema}{A numeric value of either 2012 or 2007 specifying the io schema year.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A data frame contains domestic 2 region ICFs.
}
\description{
Generate domestic 2 region inter-regional commodity flows (ICFs) table.
}
