% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions_2r.R
\name{generateTwoRegionDomesticUse}
\alias{generateTwoRegionDomesticUse}
\title{Generate two-region (SoI and RoUS) domestic use tables.}
\usage{
generateTwoRegionDomesticUse(state, year, ioschema, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{ioschema}{A numeric value of either 2012 or 2007 specifying the io schema year.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A list of domestic 2-region use tables.
}
\description{
Generate two-region (SoI and RoUS) domestic use tables.
}
