% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{generateUSDomesticUse}
\alias{generateUSDomesticUse}
\title{Generate US domestic Use table by adjusting US Use table based on Import matrix.}
\usage{
generateUSDomesticUse(iolevel, year)
}
\arguments{
\item{iolevel}{Level of detail, can be "Sector", "Summary, "Detail".}

\item{year}{A numeric value specifying the year of interest.}
}
\value{
A US Domestic Use table with rows as commodity codes and columns as industry and final demand codes
}
\description{
Generate US domestic Use table by adjusting US Use table based on Import matrix.
}
