% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{getAgFisheryForestryCommodityOutput}
\alias{getAgFisheryForestryCommodityOutput}
\title{Estimate state Ag, Fishery and Forestry commodity output ratios}
\usage{
getAgFisheryForestryCommodityOutput(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state Ag, Fishery and Forestry commodity output
for specified state with row names being BEA sector code.
}
\description{
Estimate state Ag, Fishery and Forestry commodity output ratios
}
