% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getBEASectorCodeLocation}
\alias{getBEASectorCodeLocation}
\title{Combine sector code and location to the form of code/location.}
\usage{
getBEASectorCodeLocation(sector_type, location, iolevel, specs, disagg = NULL)
}
\arguments{
\item{sector_type}{A text value specifying desired sector type,
can be "Commodity", "Industry", "FinalDemand", or "ValueAdded".}

\item{location}{A text value specifying desired location,
can be state name like "Georgia" or "RoUS" representing Rest of US.}

\item{iolevel}{Level of detail, can be "Sector", "Summary, "Detail".}

\item{specs}{A list of model specs including 'BaseIOSchema'}

\item{disagg}{optional, disaggregation specs}
}
\value{
A text value in the format of code/location.
}
\description{
Combine sector code and location to the form of code/location.
}
