% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getCountyFIPS}
\alias{getCountyFIPS}
\title{getCountyFIPS (MODIFIED)}
\usage{
getCountyFIPS(state)
}
\arguments{
\item{state}{A string character specifying the state of interest 'Georgia'}
}
\value{
A data frame contains all 159 names and FIPS for all counties in specified state
}
\description{
This function is to return a dataframe containing name and fips of each county in 
selected state in support of later data wrangling operations
}
