% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getCrossWalk}
\alias{getCrossWalk}
\title{getCrossWalk (MODIFIED)}
\usage{
getCrossWalk(start, end)
}
\arguments{
\item{start}{A string character specifying the start point}

\item{end}{A string character specifying the end point}
}
\value{
A data frame contains the croswalk table
}
\description{
This function is to return a dataframe containing the crosswalk of two different industry
classification system ('bea_sector', 'bea_summary', 'bea_detail', 'naics2007', 'naics2012', 'naics2017')
}
