% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getFileUpdateTimefromDataCommons}
\alias{getFileUpdateTimefromDataCommons}
\title{Get a datetime object for desired data file on the DataCommons server.}
\usage{
getFileUpdateTimefromDataCommons(year, iolevel, dataname)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}

\item{dataname}{Name of desired IO data, can be "Make", "Use", "DomesticUse", "CommodityOutput, and "IndustryOutput".}
}
\value{
A datetime object for desired data file on the DataCommons server.
}
\description{
Get a datetime object for desired data file on the DataCommons server.
}
