% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getFinalDemandCodes}
\alias{getFinalDemandCodes}
\title{Get codes of final demand.}
\usage{
getFinalDemandCodes(iolevel, specs)
}
\arguments{
\item{iolevel}{Level of detail, can be "Sector", "Summary, "Detail".}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A vector of final demand codes.
}
\description{
Get codes of final demand.
}
