% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{getNationalMake}
\alias{getNationalMake}
\title{Get US Make table of specified iolevel and year.}
\usage{
getNationalMake(iolevel, year)
}
\arguments{
\item{iolevel}{Level of detail, can be "Sector", "Summary, "Detail".}

\item{year}{A numeric value specifying the year of interest.}
}
\value{
The US make table of specified iolevel and year.
}
\description{
Get US Make table of specified iolevel and year.
}
