% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{getNationalUse}
\alias{getNationalUse}
\title{Get US Use table (intermediaete + final demand) of specified iolevel and year.}
\usage{
getNationalUse(iolevel, year)
}
\arguments{
\item{iolevel}{BEA sector level of detail, currently can only be "Summary",
theoretically can be "Detail", or "Sector" in future versions.}

\item{year}{A numeric value specifying the year of interest.}
}
\value{
The US Use table (intermediaete + final demand) of specified iolevel and year.
}
\description{
Get US Use table (intermediaete + final demand) of specified iolevel and year.
}
