% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getRegistryonDataCommons}
\alias{getRegistryonDataCommons}
\title{Get data registry on Data Commons.}
\usage{
getRegistryonDataCommons(data_group = "stateio")
}
\arguments{
\item{data_group}{A string specifying name of data group that is used as the
subdirectory in Data Commons, can be "stateio", "flowsa/FlowBySector", or
"flowsa/FlowByActivity".}
}
\value{
A dataframe of StateIO data registry on Data Commons.
}
\description{
Get data registry on Data Commons.
}
