% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getStateAbbreviation}
\alias{getStateAbbreviation}
\title{This function converts US state name, for example "Alabama",
to a two-character state abbreviation "AL". Can take "District of Columbia".}
\usage{
getStateAbbreviation(state)
}
\arguments{
\item{state}{A string character specifying the full name of a US state.}
}
\value{
two-character abbreviation of a US state.
}
\description{
This function converts US state name, for example "Alabama",
to a two-character state abbreviation "AL". Can take "District of Columbia".
}
