% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{getStateCommodityOutputRatioEstimates}
\alias{getStateCommodityOutputRatioEstimates}
\title{Estimate state commodity output from alternative sources, calculate ratios.}
\usage{
getStateCommodityOutputRatioEstimates(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state commodity output for specified state with row names being BEA sector code.
}
\description{
Estimate state commodity output from alternative sources, calculate ratios.
}
