% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{getStateCompensationbyBEASummary}
\alias{getStateCompensationbyBEASummary}
\title{Load BEA State Compensation data from pre-saved .rds files.
Map to BEA Summary sectors.}
\usage{
getStateCompensationbyBEASummary(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema',}
}
\value{
A data frame contains State Compensation by BEA Summary.
}
\description{
Load BEA State Compensation data from pre-saved .rds files.
Map to BEA Summary sectors.
}
