% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{getStateEmpCompensation}
\alias{getStateEmpCompensation}
\title{Get industry-level Compensation for all states at a specific year.}
\usage{
getStateEmpCompensation(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state Compensation for all states at a specific year.
}
\description{
Get industry-level Compensation for all states at a specific year.
}
