% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{getStateEmploymentbyBEASummary}
\alias{getStateEmploymentbyBEASummary}
\title{Load BEA State Employment data from pre-saved .rds files and State Employment
FlowBySector data from flowsa.
Map to BEA Summary sectors.}
\usage{
getStateEmploymentbyBEASummary(year, specs, datasource = "BEA")
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema',}

\item{datasource}{A str to differentiate primary data source for employment}
}
\value{
A data frame contains State Employment by BEA Summary.
}
\description{
Load BEA State Employment data from pre-saved .rds files and State Employment
FlowBySector data from flowsa.
Map to BEA Summary sectors.
}
