% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getStateIOData}
\alias{getStateIOData}
\title{Load state IO data from Data Commons}
\usage{
getStateIOData(dataname, year, version)
}
\arguments{
\item{dataname}{A string specifying data name, can be "State_Summary_Use".}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{version}{A string specifying version of the data, can be "v0.1.0".}
}
\value{
A data frame contains state data from FLOWSA.
}
\description{
Load state IO data from Data Commons
}
