% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getStateIODataRegistryonDataCommons}
\alias{getStateIODataRegistryonDataCommons}
\title{Get StateIO data registry on Data Commons.}
\usage{
getStateIODataRegistryonDataCommons()
}
\value{
A dataframe of StateIO data registry on Data Commons.
}
\description{
Get StateIO data registry on Data Commons.
}
