% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{getStateModelDisaggSpecs}
\alias{getStateModelDisaggSpecs}
\title{Read and assign disaggregation specifications}
\usage{
getStateModelDisaggSpecs()
}
\value{
A stateior model object with the disaggregation specs loaded.
}
\description{
Read and assign disaggregation specifications
}
