% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{getStateSpecificDisaggSpecs}
\alias{getStateSpecificDisaggSpecs}
\title{Read in state-specific disaggregation values
This function assumes values contained in statefile will modify disaggregation
values present in the main disaggregation config file.}
\usage{
getStateSpecificDisaggSpecs(disaggConfigpath, statefile)
}
\arguments{
\item{disaggConfigpath}{str, path for statefile}

\item{statefile}{str, name of state-specific disaggregation spec file that will
modify the standard configfile for each state. Should be 1 config file that has
the required modifications of configfile for each state}
}
\value{
A stateior model object with the state-specific disaggregation specs
included in model$specs$STateDisaggSpecs object
}
\description{
Read in state-specific disaggregation values
This function assumes values contained in statefile will modify disaggregation
values present in the main disaggregation config file.
}
