% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{getStateTax}
\alias{getStateTax}
\title{Get industry-level Tax for all states at a specific year.}
\usage{
getStateTax(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state Tax for all states at a specific year.
}
\description{
Get industry-level Tax for all states at a specific year.
}
