% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{getTwoRegionCommodityOutput}
\alias{getTwoRegionCommodityOutput}
\title{Load commodity output tables of a state of interest (SoI)
and its corresponding rest-of-US (RoUS) for a given year.}
\usage{
getTwoRegionCommodityOutput(state, year, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A data.frame of SoI's and RoUS' commodity output.
}
\description{
Load commodity output tables of a SoI and its corresponding RoUS
for a given year.
}
