% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getTwoRegionDataFileName}
\alias{getTwoRegionDataFileName}
\title{Generate two-region data filename with .rda as suffix.}
\usage{
getTwoRegionDataFileName(year, iolevel, dataname)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}

\item{dataname}{Name of desired IO data, can be "Make", "Use", "DomesticUse", "CommodityOutput, and "IndustryOutput".}
}
\value{
A string of two-region data filename with .rda as suffix.
}
\description{
Generate two-region data filename with .rda as suffix.
}
