% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{getTwoRegionDomesticUseTransactions}
\alias{getTwoRegionDomesticUseTransactions}
\title{Load domestic use transactions (intermediate consumption)
of a state of interest (SoI) and its corresponding rest-of-US (RoUS)
for a given year.}
\usage{
getTwoRegionDomesticUseTransactions(state, year, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A data.frame of SoI's and RoUS' domestic use transactions.
}
\description{
Load domestic use transactions (intermediate consumption)
of a SoI and its corresponding RoUS for a given year.
}
