% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{GovConsumption_2007_2019}
\alias{GovConsumption_2007_2019}
\title{BEA US Gov Consumption table (2007-2019)}
\format{
A data frame with 60 obs. and 15 variables
\describe{
 \item{Line}{numeric value}
 \item{Description}{text code}
 \item{2007}{current year USD}
 \item{2008}{current year USD}
 \item{2009}{current year USD}
 \item{2010}{current year USD}
 \item{2011}{current year USD}
 \item{2012}{current year USD}
 \item{2013}{current year USD}
 \item{2014}{current year USD}
 \item{2015}{current year USD}
 \item{2016}{current year USD}
 \item{2017}{current year USD}
}
}
\source{
\url{https://apps.bea.gov/national/Release/XLS/Survey/Section3All_xls.xlsx}
}
\usage{
GovConsumption_2007_2019
}
\description{
BEA US Gov Consumption table (2007-2019)
}
\keyword{datasets}
