% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{loadDataCommonsfile}
\alias{loadDataCommonsfile}
\title{Load flowsa FlowByActivity or FlowBySector data from Data Commons}
\usage{
loadDataCommonsfile(dataname, year)
}
\arguments{
\item{dataname}{A string specifying data name, can be "NOAA_FisheryLandings".}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state data from FLOWSA.
}
\description{
Load flowsa FlowByActivity or FlowBySector data from Data Commons
}
