% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{loadDatafromFLOWSA}
\alias{loadDatafromFLOWSA}
\title{Load flowbyactivity data from FLOWSA}
\usage{
loadDatafromFLOWSA(flowclass, year, datasource, geographic_level = NULL)
}
\arguments{
\item{flowclass}{A character value specifying flow class, can be "Money".}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{datasource}{A character value specifying data source.}
}
\value{
A data frame contains state data from FLOWSA.
}
\description{
Load flowbyactivity data from FLOWSA
}
