% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{loadStateDomesticUseTable}
\alias{loadStateDomesticUseTable}
\title{Load domestic use table of a specific state for a given year.}
\usage{
loadStateDomesticUseTable(state, year, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A dataframe of state domestic use table.
}
\description{
Load domestic use table of a specific state for a given year.
}
