% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{loadStateIndustryOutput}
\alias{loadStateIndustryOutput}
\title{Load industry output of a specific state for a given year.}
\usage{
loadStateIndustryOutput(state, year, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A dataframe of state industry output.
}
\description{
Load industry output of a specific state for a given year.
}
