% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{loadStateIODataFile}
\alias{loadStateIODataFile}
\title{Load StateIO data file from Data Commons or local data directory.}
\usage{
loadStateIODataFile(filename, ver = NULL)
}
\arguments{
\item{filename}{A string specifying filename, e.g. "State_Summary_Use_2017".}

\item{ver}{A string specifying version of the data, default is NULL, can be "v0.1.0".}
}
\value{
A StateIO data product (usually a list of dataframes).
}
\description{
If the file is not found locally, it will be downloaded from Data Commons. Available
files can be found on https://dmap-data-commons-ord.s3.amazonaws.com/index.html#stateio/.
These include Use, DomesticUse, and IndustryOutput tables, among others.
}
