% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{loadStateMakeTable}
\alias{loadStateMakeTable}
\title{Load make table of a specific state for a given year.}
\usage{
loadStateMakeTable(state, year, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A dataframe of state make table.
}
\description{
Load make table of a specific state for a given year.
}
