% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{loadStateUseTable}
\alias{loadStateUseTable}
\title{Load use table of a specific state for a given year.}
\usage{
loadStateUseTable(state, year, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A dataframe of state use table.
}
\description{
Load use table of a specific state for a given year.
}
