% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{loadTwoRegionDomesticUseTable}
\alias{loadTwoRegionDomesticUseTable}
\title{Load domestic use tables of a state of interest (SoI) and its corresponding rest-of-US (RoUS) for a given year.}
\usage{
loadTwoRegionDomesticUseTable(state, year, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A dataframe of a state of interest (SoI) and its corresponding rest-of-US (RoUS) domestic use tables.
}
\description{
Load domestic use tables of a state of interest (SoI) and its corresponding rest-of-US (RoUS) for a given year.
}
