% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{loadTwoRegionIndustryOutput}
\alias{loadTwoRegionIndustryOutput}
\title{Load industry output tables of a state of interest (SoI) and its corresponding rest-of-US (RoUS) for a given year.}
\usage{
loadTwoRegionIndustryOutput(state, year, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A dataframe of a state of interest (SoI) and its corresponding rest-of-US (RoUS) industry output.
}
\description{
Load industry output tables of a state of interest (SoI) and its corresponding rest-of-US (RoUS) for a given year.
}
