% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{loadTwoRegionIOData}
\alias{loadTwoRegionIOData}
\title{Load two-region IO data of specified iolevel and year
from user's local directory or the EPA Data Commons.}
\usage{
loadTwoRegionIOData(year, iolevel, dataname)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, currently can only be "Summary",
theoretically can be "Detail", or "Sector" in future versions.}

\item{dataname}{Name of desired IO data, can be "Make", "Use", "DomesticUse",
"UseTransactions", "FinalDemand", "DomesticUseTransactions", "DomesticFinalDemand",
"CommodityOutput, and "IndustryOutput", "Demand", and "CompleteDemand".}
}
\value{
A list of two-region IO data of specified iolevel and year.
}
\description{
Load two-region IO data of specified iolevel and year
from user's local directory or the EPA Data Commons.
}
