% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectorMappingFunctions.R
\name{mapBLSQCEWtoBEA}
\alias{mapBLSQCEWtoBEA}
\title{Map BLS QCEW data from NAICS to BEA (Sector/Summary/Detail) sectors.}
\usage{
mapBLSQCEWtoBEA(bls_qcew, year, iolevel)
}
\arguments{
\item{bls_qcew}{A data frame contains BLS QCEW by NAICS data, directly loaded from flowsa.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{A character value, the level of BEA sector to map to.}
}
\value{
A data frame contains BLS QCEW by BEA.
}
\description{
Map BLS QCEW data from NAICS to BEA (Sector/Summary/Detail) sectors.
}
