% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{mapStateTabletoBEASummary}
\alias{mapStateTabletoBEASummary}
\title{Map state table to BEA Summary, mark sectors that need allocation}
\usage{
mapStateTabletoBEASummary(statetablename, year)
}
\arguments{
\item{statetablename}{Name of pre-saved state table,
can be GVA, Tax, Employment Compensation, and GOS.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state value for all states with row names being BEA sector code.
}
\description{
Map state table to BEA Summary, mark sectors that need allocation
}
