% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{plotICFandRPC}
\alias{plotICFandRPC}
\title{Plot SoI2SoI or RoUS2RoUS ICF and state RPC}
\usage{
plotICFandRPC(ICF_df, RPC_df)
}
\arguments{
\item{ICF_df}{A data.frame contains SoI2SoI ICF ratios}

\item{RPC_df}{A data.frame contains state RPC ratios}
}
\description{
Plot SoI2SoI or RoUS2RoUS ICF and state RPC
}
