% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{plotResidual}
\alias{plotResidual}
\title{Plot SoI2SoI ICF and state RPC}
\usage{
plotResidual(df)
}
\arguments{
\item{df}{A data.frame contains residual values}
}
\description{
Plot SoI2SoI ICF and state RPC
}
