% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationFunctions.R
\name{reportValidationResult}
\alias{reportValidationResult}
\title{Report failures from validation in printed messages}
\usage{
reportValidationResult(failures)
}
\arguments{
\item{failures}{A data frame or a vector of failures}
}
\description{
Report failures from validation in printed messages
}
