% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDisaggFunctions.R
\name{splitFullUse}
\alias{splitFullUse}
\title{Separate full use table into model components}
\usage{
splitFullUse(model, domestic = FALSE)
}
\arguments{
\item{model}{An stateior model object with model specs and specific IO tables loaded}

\item{domestic}{A boolean that indicates whether the table to format is the domesticUse table or not}
}
\value{
A model object with FullUse split into UseTransactions, FinalDemand, and UseValueAdded objects
}
\description{
Separate full use table into model components
}
