% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{State_Compensation_2007_2017}
\alias{State_Compensation_2007_2017}
\title{BEA State Employee Compensation table (2007-2017)}
\format{
A data frame with 4736 obs. and 14 variables
\describe{
 \item{GeoName}{text code}
 \item{LineCode}{integer}
 \item{Description}{text code}
 \item{2007}{current year USD}
 \item{2008}{current year USD}
 \item{2009}{current year USD}
 \item{2010}{current year USD}
 \item{2011}{current year USD}
 \item{2012}{current year USD}
 \item{2013}{current year USD}
 \item{2014}{current year USD}
 \item{2015}{current year USD}
 \item{2016}{current year USD}
 \item{2017}{current year USD}
}
}
\source{
\url{https://apps.bea.gov/regional/zip/SAGDP.zip}
}
\usage{
State_Compensation_2007_2017
}
\description{
BEA State Employee Compensation table (2007-2017)
}
\keyword{datasets}
