% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{State_Employment_2009_2018}
\alias{State_Employment_2009_2018}
\title{BEA State Employment table (2009-2018)}
\format{
A data frame with 7080 obs. and 13 variables
\describe{
 \item{GeoFips}{text code}
 \item{GeoName}{text code}
 \item{LineCode}{integer}
 \item{2009}{number of employee}
 \item{2010}{number of employee}
 \item{2011}{number of employee}
 \item{2012}{number of employee}
 \item{2013}{number of employee}
 \item{2014}{number of employee}
 \item{2015}{number of employee}
 \item{2016}{number of employee}
 \item{2017}{number of employee}
 \item{2018}{number of employee}
}
}
\source{
\url{https://apps.bea.gov/API/signup/index.cfm}
}
\usage{
State_Employment_2009_2018
}
\description{
BEA State Employment table (2009-2018)
}
\keyword{datasets}
