% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{State_GrossValueAdded_2012}
\alias{State_GrossValueAdded_2012}
\title{State gross value addaed 2012}
\format{
A dataframe with 156 obs. (representing Summary GVA sectors in 52 states/regions)
and 71 variables (representing Summary industries)
}
\usage{
State_GrossValueAdded_2012
}
\description{
State gross value addaed 2012
}
\keyword{datasets}
