% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{State_PCE_2007_2018}
\alias{State_PCE_2007_2018}
\title{BEA State PCE table (2007-2018)}
\format{
A data frame with 1224 obs. and 15 variables
\describe{
 \item{GeoName}{text code}
 \item{Line}{integer}
 \item{Description}{text code}
 \item{2007}{current year USD}
 \item{2008}{current year USD}
 \item{2009}{current year USD}
 \item{2010}{current year USD}
 \item{2011}{current year USD}
 \item{2012}{current year USD}
 \item{2013}{current year USD}
 \item{2014}{current year USD}
 \item{2015}{current year USD}
 \item{2016}{current year USD}
 \item{2017}{current year USD}
 \item{2018}{current year USD}
}
}
\source{
\url{https://apps.bea.gov/regional/zip/SAEXP.zip}
}
\usage{
State_PCE_2007_2018
}
\description{
BEA State PCE table (2007-2018)
}
\keyword{datasets}
