% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{State_Summary_Make_2012}
\alias{State_Summary_Make_2012}
\title{State Summary 2012 Make (in a list format)}
\format{
A list with 52 elements:
each element is a dataframe with 71 obs. (representing Summary industries)
and 73 variables (representing Summary commodities)
}
\usage{
State_Summary_Make_2012
}
\description{
State Summary 2012 Make (in a list format)
}
\keyword{datasets}
