% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{State_Summary_Use_2012}
\alias{State_Summary_Use_2012}
\title{State Summary 2012 Use (in a list format)}
\format{
A list with 52 elements:
each element is a dataframe with 76 obs. (representing Summary commodities and value added)
and 91 variables (representing Summary industries and final demand sectors)
}
\usage{
State_Summary_Use_2012
}
\description{
State Summary 2012 Use (in a list format)
}
\keyword{datasets}
