% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{StateUS_VA_Ratio_2012}
\alias{StateUS_VA_Ratio_2012}
\title{State-US value addaed ratio 2012}
\format{
A dataframe with 3692 obs. and 3 variables
\describe{
 \item{BEA_2012_Summary_Code}{character}
 \item{GeoName}{character}
 \item{Ratio}{numeric}
}
}
\usage{
StateUS_VA_Ratio_2012
}
\description{
State-US value addaed ratio 2012
}
\keyword{datasets}
