% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{TwoRegion_Summary_CommodityOutput_2012}
\alias{TwoRegion_Summary_CommodityOutput_2012}
\title{Two-region Summary 2012 Commodity Output (in a list format)}
\format{
A list with 51 elements, representing 50 states and DC:
each element is a dataframe with 73 obs. (representing Summary commodities)
and 2 variable (SoI and RoUS Commodity Output)
}
\usage{
TwoRegion_Summary_CommodityOutput_2012
}
\description{
Two-region Summary 2012 Commodity Output (in a list format)
}
\keyword{datasets}
