% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{TwoRegion_Summary_Use_2012}
\alias{TwoRegion_Summary_Use_2012}
\title{Two-region Summary 2012 Use (in a list format)}
\format{
A list with 51 elements, representing 50 states and DC:
each element is a dataframe with 152 obs. (representing Summary SoI and RoUS commodities and value added)
and 182 variables (representing Summary SoI and RoUS industries and final demand sectors)
}
\usage{
TwoRegion_Summary_Use_2012
}
\description{
Two-region Summary 2012 Use (in a list format)
}
\keyword{datasets}
