% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationFunctions.R
\name{validateStateUseAgainstNationlUse}
\alias{validateStateUseAgainstNationlUse}
\title{Validate state Use against US Use, cell-by-cell}
\usage{
validateStateUseAgainstNationlUse(domestic = FALSE, rel_diff = FALSE)
}
\arguments{
\item{domestic}{A logical value indicating whether to compare Domestic Use or not}

\item{rel_diff}{A logical value indicating whether to validate relative difference or not}
}
\value{
A data.frame or vector contains failures
}
\description{
Validate state Use against US Use, cell-by-cell
}
