% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{writeDatafileMeta}
\alias{writeDatafileMeta}
\title{Write a datetime object for desired data file to local folder.}
\usage{
writeDatafileMeta(year, iolevel, dataname, path)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, currently can only be "Summary",
theoretically can be "Detail", or "Sector" in future versions.}

\item{dataname}{Name of desired IO data, can be "Make", "Use", "DomesticUse",
"CommodityOutput, and "IndustryOutput".}

\item{path}{User-defined local path.}
}
\value{
A datetime object for desired data file to local folder.
}
\description{
Get a datetime object for desired data file to local folder.
}
