% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeStateDemandModel}
\alias{writeStateDemandModel}
\title{Write state demand model to .rda}
\usage{
writeStateDemandModel(model)
}
\arguments{
\item{model}{A state demand model, must contain a list of state Use tables and
a list of state domestic Use tables.}
}
\description{
Write state demand model to .rda
}
