% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeStateIODatatoCSV}
\alias{writeStateIODatatoCSV}
\title{Write StateIO data to specified format, .csv or .xlsx, to output folder.
The data are loaded from Data Commons or local data directory.}
\usage{
writeStateIODatatoCSV(filename, ver = NULL, state, outputfolder)
}
\arguments{
\item{filename}{A string specifying filename, e.g. "State_Summary_Use_2017".}

\item{ver}{A string specifying version of the data, default is NULL, can be "v0.1.0".}

\item{state}{State name.}

\item{outputfolder}{A directory to write matrices out to.}
}
\description{
Writes StateIO data as .csv or .xlsx files to output folder.
}
