% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeStateSupplyModel}
\alias{writeStateSupplyModel}
\title{Write state supply model to .rda}
\usage{
writeStateSupplyModel(model)
}
\arguments{
\item{model}{A state supply model, must contain a list of state Make tables,
a list of state industry output, and a list of state commodity output.}
}
\description{
Write state supply model to .rda
}
