% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{allocateStateTabletoBEASummary}
\alias{allocateStateTabletoBEASummary}
\title{Allocate state table (GVA, Tax, Employment Compensation, and GOS)
to BEA Summary based on specified weight}
\usage{
allocateStateTabletoBEASummary(
  statetablename,
  year,
  allocationweightsource,
  specs
)
}
\arguments{
\item{statetablename}{Name of pre-saved state table,
can be GVA, Tax, Employment Compensation, and GOS.}

\item{year}{A numeric value specifying the year of interest.}

\item{allocationweightsource}{Source of allocation weight, can be "Employment"
or "Compensation".}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains allocated state value
for all states with row names being BEA sector code.
}
\description{
Allocate state table (GVA, Tax, Employment Compensation, and GOS)
to BEA Summary based on specified weight
}
