% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{assembleStateSummaryGrossValueAdded}
\alias{assembleStateSummaryGrossValueAdded}
\title{Assemble Summary-level gross value added sectors (V001, V002, V003) for all states at a specific year.}
\usage{
assembleStateSummaryGrossValueAdded(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A data frame contains Summary-level gross value added (V001, V002, V003) for all states at a specific year.
}
\description{
Assemble Summary-level gross value added sectors (V001, V002, V003) for all states at a specific year.
}
