% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{assembleTwoRegionIO}
\alias{assembleTwoRegionIO}
\title{Assemble two-region make, use, domestic use, and Use tables as well as commodity and industry outputs.}
\usage{
assembleTwoRegionIO(year, iolevel)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A list of two-region make, use, domestic use, and Use tables
as well as commodity and industry outputs by state.
}
\description{
Assemble two-region make and use tables as well as commodity and industry outputs.
}
