% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommodityFlowFunctions.R
\name{calculateCommodityFlowRatios}
\alias{calculateCommodityFlowRatios}
\title{Calculate domestic/import/export commodity flow ratios by state}
\usage{
calculateCommodityFlowRatios(state, year, flow_ratio_type, ioschema, iolevel)
}
\arguments{
\item{state}{State name.}

\item{year}{A numeric value between 2012 and 2017 specifying the year of interest.}

\item{flow_ratio_type}{Type of commodity flow, can be "domestic", "export", or "import".}

\item{ioschema}{A numeric value of either 2012 or 2007 specifying the io schema year.}

\item{iolevel}{BEA sector level of detail, currently can only be "Summary",
theoretically can be "Detail", or "Sector" in future versions.}
}
\value{
A data frame contains commodity flow ratios by BEA.
}
\description{
Calculate domestic/import/export commodity flow ratios by state
}
