% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommodityFlowFunctions.R
\name{calculateElectricityFlowRatios}
\alias{calculateElectricityFlowRatios}
\title{Calculate domestic electricity flow ratios by state}
\usage{
calculateElectricityFlowRatios(state, year)
}
\arguments{
\item{state}{State name.}

\item{year}{A numeric value between 2012 and 2017 specifying the year of interest.}
}
\value{
A data frame contains electricity flow ratios by BEA.
}
\description{
Calculate domestic electricity flow ratios by state
}
