% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteregionalCommodityFlowFunctions.R
\name{calculateLocalandTradedRatios}
\alias{calculateLocalandTradedRatios}
\title{Calculate domestic local and traded ratio, which will be used in the next function generateDomestic2RegionICFs}
\usage{
calculateLocalandTradedRatios(state, year, SoI = TRUE, ioschema, iolevel)
}
\arguments{
\item{state}{State name.}

\item{year}{A numeric value between 2012 and 2017 specifying the year of interest.}

\item{SoI}{A boolean variable indicating whether to calculate local and traded ratios for SoI or RoU.}

\item{ioschema}{A numeric value of either 2012 or 2007 specifying the io schema year.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A data frame contains local and traded ratios by BEA sectors for the specified state.
}
\description{
Calculate domestic local and traded ratio, which will be used in the next function generateDomestic2RegionICFs
}
