% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteregionalCommodityFlowFunctions.R
\name{calculateLocalandTradedRatios}
\alias{calculateLocalandTradedRatios}
\title{Calculate domestic local and traded ratio, used for calculating ICF ratios.}
\usage{
calculateLocalandTradedRatios(state, year, SoI = TRUE, specs, iolevel)
}
\arguments{
\item{state}{State name.}

\item{year}{A numeric value specifying the year of interest.}

\item{SoI}{A logical variable indicating whether to calculate
local and traded ratios for SoI or RoUS.}

\item{specs}{A list of model specs including 'BaseIOSchema'}

\item{iolevel}{BEA sector level of detail, currently can only be "Summary",
theoretically can be "Detail", or "Sector" in future versions.}
}
\value{
A data frame contains local and traded ratios by BEA sectors
for the specified state.
}
\description{
Calculate domestic local and traded ratio, used for calculating ICF ratios.
}
