% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{calculateStateCommodityOutputRatio}
\alias{calculateStateCommodityOutputRatio}
\title{Calculate state-US Commodity Output ratios at BEA Summary level.
Apply row sum to state and US Make tables to get commodity output vectors and
then Commodity Output Ratio (COR).}
\usage{
calculateStateCommodityOutputRatio(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains state-US Commodity Output ratios at BEA Summary level.
}
\description{
Calculate state-US Commodity Output ratios at BEA Summary level.
Apply row sum to state and US Make tables to get commodity output vectors and
then Commodity Output Ratio (COR).
}
