% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{calculateStateFedGovExpenditureRatio}
\alias{calculateStateFedGovExpenditureRatio}
\title{Calculate state fed government expenditure ratio at BEA Summary level.}
\usage{
calculateStateFedGovExpenditureRatio(year)
}
\arguments{
\item{year}{A numeric value between 2008 and 2019 specifying the year of interest.}
}
\value{
A data frame contains state fed government expenditure ratio for all states at a specific year at BEA Summary level.
}
\description{
Calculate state fed government expenditure ratio at BEA Summary level.
}
