% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{calculateStateSLGovExpenditureRatio}
\alias{calculateStateSLGovExpenditureRatio}
\title{Calculate state S&L government expenditure ratio at BEA Summary level.}
\usage{
calculateStateSLGovExpenditureRatio(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state S&L government expenditure ratio for all states at a specific year at BEA Summary level.
}
\description{
Calculate state S&L government expenditure ratio at BEA Summary level.
}
