% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{calculateStatetoBEASummaryAllocationFactor}
\alias{calculateStatetoBEASummaryAllocationFactor}
\title{Calculate allocation factors based on state-level data, such as employment}
\usage{
calculateStatetoBEASummaryAllocationFactor(year, allocationweightsource)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{allocationweightsource}{A string specifying the source being used
as the weight in the allocation.}
}
\value{
A data frame contains allocation factors
for all states with row names being BEA sector code.
}
\description{
Calculate allocation factors based on state-level data, such as employment
}
