% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{calculateStateUSEmpCompensationRatio}
\alias{calculateStateUSEmpCompensationRatio}
\title{Calculate state-US employee compensation ratios at BEA Summary level.}
\usage{
calculateStateUSEmpCompensationRatio(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state-US employment compensation ratios for all states at a specific year at BEA Summary level.
}
\description{
Calculate state-US employee compensation ratios at BEA Summary level.
}
