% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{calculateStateUSVARatiobyLineCode}
\alias{calculateStateUSVARatiobyLineCode}
\title{Calculate state-US GVA (value added) ratios by BEA State LineCode.}
\usage{
calculateStateUSVARatiobyLineCode(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains ratios of state/US GVA (value added) for all states at a specific year by BEA State LineCode.
}
\description{
Calculate state-US GVA (value added) ratios by BEA State LineCode.
}
