% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommodityFlowFunctions.R
\name{calculateWasteManagementServiceFlowRatios}
\alias{calculateWasteManagementServiceFlowRatios}
\title{Calculate domestic waste management services flow ratios by state}
\usage{
calculateWasteManagementServiceFlowRatios(state, year)
}
\arguments{
\item{state}{State name.}

\item{year}{A numeric value between 2012 and 2017 specifying the year of interest.}
}
\value{
A data frame contains waste management services flow ratios by BEA.
}
\description{
Calculate domestic waste management services flow ratios by state
}
