% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{checkFileonDataCommons}
\alias{checkFileonDataCommons}
\title{Check if file is available on Data Commons. Stop function execution if not.}
\usage{
checkFileonDataCommons(file)
}
\arguments{
\item{file}{A string specifying file, e.g. "State_Summary_Use_2017_v0.1.0_rds".}
}
\description{
Check if file is available on Data Commons. Stop function execution if not.
}
