% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{estimateStateExport}
\alias{estimateStateExport}
\title{Estimate state export at BEA Summary level.}
\usage{
estimateStateExport(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains state export for all states at a specific year at BEA Summary level.
}
\description{
Estimate state export at BEA Summary level.
}
