% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{estimateStateFedGovExpenditure}
\alias{estimateStateFedGovExpenditure}
\title{Estimate state fed government expenditure at BEA Summary level.}
\usage{
estimateStateFedGovExpenditure(year, specs)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains state fed government expenditure for all states
at a specific year at BEA Summary level.
}
\description{
Estimate state fed government expenditure at BEA Summary level.
}
