% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{estimateStateHouseholdDemand}
\alias{estimateStateHouseholdDemand}
\title{Estimate state household demand at BEA Summary level.}
\usage{
estimateStateHouseholdDemand(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains state household demand for all states at a specific year at BEA Summary level.
}
\description{
Estimate state household demand at BEA Summary level.
}
