% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{estimateStatePrivateInvestment}
\alias{estimateStatePrivateInvestment}
\title{Estimate state private investment at BEA Summary level.
Apply state PCE ratio to F02R.
Apply state Gross Output ratio to F02S, F02E, F02N, and F030.}
\usage{
estimateStatePrivateInvestment(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state household demand for all states at a specific year at BEA Summary level.
}
\description{
Estimate state private investment at BEA Summary level.
Apply state PCE ratio to F02R.
Apply state Gross Output ratio to F02S, F02E, F02N, and F030.
}
