% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{estimateStateSLGovExpenditure}
\alias{estimateStateSLGovExpenditure}
\title{Estimate state S&L government expenditure at BEA Summary level.}
\usage{
estimateStateSLGovExpenditure(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state S&L government expenditure for all states at a specific year at BEA Summary level.
}
\description{
Estimate state S&L government expenditure at BEA Summary level.
}
