% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{finalizeStateUSValueAddedRatio}
\alias{finalizeStateUSValueAddedRatio}
\title{Finalize state-US value added ratios using RAS balancing method}
\usage{
finalizeStateUSValueAddedRatio(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains RAS-balanced state-US value added ratios
0 - Define BEA_col and year_col
1 - Load state GVA/value added (VA) by industry for the given year.
Map to BEA Summary level
2 - Where VA must be allocated from a LineCode to BEA Summary industries,
calculate allocation factors using specified allocationweightsource.
When using state employment (from BEA) as source for allocation,
introduce national GVA to disaggregate state employment
in real estate and gov industries from LineCode to BEA Summary.
3 - Load US Summary Make table for given year
Generate US Summary Make Transaction and Industry and Commodity Output
4 - Calculate state_US_VA_ratio
For each state and each industry, divide state VA by US VA.
For each state, estimate industry output based on US industry output and
state_US_VA_ratio.
For industries whose state_US_VA_ratio were calculatedbased on disaggregated
state VA, apply RAS to balance state industry output and state_US_VA_ratio.
Use the adjusted state_US_VA_ratio to calculate
state make transaction, industry and commodity output.
}
\description{
Finalize state-US value added ratios using RAS balancing method
}
