% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteregionalCommodityFlowFunctions.R
\name{generateDomestic2RegionICFs}
\alias{generateDomestic2RegionICFs}
\title{Generate domestic 2 region inter-regional commodity flows (ICFs) table.}
\usage{
generateDomestic2RegionICFs(
  state,
  year,
  ioschema,
  iolevel,
  ICF_sensitivity_analysis = FALSE,
  adjust_by = 0
)
}
\arguments{
\item{state}{State name.}

\item{year}{A numeric value between 2012 and 2017 specifying the year of interest.}

\item{ioschema}{A numeric value of either 2012 or 2007 specifying the io schema year.}

\item{iolevel}{BEA sector level of detail, currently can only be "Summary",
theoretically can be "Detail", or "Sector" in future versions.}

\item{ICF_sensitivity_analysis}{A logical value indicating whether to conduct
sensitivity analysis on ICF, default is FALSE.}

\item{adjust_by}{A numeric value between 0 and 1 indicating the manual adjustment
to ICF if a sensitivity analysis is conducted, default is 0 due to no SA.
#' @return A data frame contains domestic 2 region ICFs.}
}
\description{
Generate domestic 2 region inter-regional commodity flows (ICFs) table.
}
