% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{getFAFCommodityOutput}
\alias{getFAFCommodityOutput}
\title{Estimate state FAF commodity output ratios}
\usage{
getFAFCommodityOutput(year, specs)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema',}
}
\value{
A data frame contains state FAF commodity output
for specified state with row names being BEA sector code.
}
\description{
Estimate state FAF commodity output ratios
}
