% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{getNationalUse}
\alias{getNationalUse}
\title{Get US Use table (intermediaete + final demand) of specified iolevel and year.}
\usage{
getNationalUse(iolevel, year)
}
\arguments{
\item{iolevel}{Level of detail, can be "Sector", "Summary, "Detail".}

\item{year}{A numeric value specifying the year of interest.
#' @return The US Use table (intermediaete + final demand) of specified iolevel and year.}
}
\description{
Get US Use table (intermediaete + final demand) of specified iolevel and year.
}
