% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{getStateEmpCompensation}
\alias{getStateEmpCompensation}
\title{Get industry-level Compensation for all states at a specific year.}
\usage{
getStateEmpCompensation(year, specs)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains state Compensation for all states at a specific year.
}
\description{
Get industry-level Compensation for all states at a specific year.
}
