% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{getStateEmploymentbyBEASummary}
\alias{getStateEmploymentbyBEASummary}
\title{Load BEA and BLS QCEW State Employment data from pre-saved .rds files and FLOWSA.
Map to BEA Summary sectors.}
\usage{
getStateEmploymentbyBEASummary(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains State Emp by BEA Summary.
}
\description{
Load BEA and BLS QCEW State Employment data from pre-saved .rds files and FLOWSA.
Map to BEA Summary sectors.
}
