% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSupplyFunctions.R
\name{getStateEmploymentTable}
\alias{getStateEmploymentTable}
\title{Get a table of state employment by BEA Summary sector.}
\usage{
getStateEmploymentTable(year)
}
\arguments{
\item{year}{A numeric value specifying the year of interest.}
}
\value{
A data frame containing employment count by state and BEA Summary sector.
}
\description{
Get a table of state employment by BEA Summary sector.
}
