% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateUseFunctions.R
\name{getStateGOS}
\alias{getStateGOS}
\title{Get industry-level Gross Operating Surplus (GOS) for all states at a specific year.}
\usage{
getStateGOS(year, specs)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A data frame contains state GOS for all states at a specific year.
}
\description{
Get industry-level Gross Operating Surplus (GOS) for all states at a specific year.
}
