% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getStateIODataRegistryonDataCommons}
\alias{getStateIODataRegistryonDataCommons}
\title{Get state IO data registry on Data Commons.}
\usage{
getStateIODataRegistryonDataCommons()
}
\value{
A dataframe of state IO data registry on Data Commons.
}
\description{
Get state IO data registry on Data Commons.
}
