% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDemandFunctions.R
\name{getStatePCE}
\alias{getStatePCE}
\title{Get commodity-level Personal Consumption Expenditure (PCE) for all states at a specific year.}
\usage{
getStatePCE(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A data frame contains state PCE for all states at a specific year.
}
\description{
Get commodity-level Personal Consumption Expenditure (PCE) for all states at a specific year.
}
