% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{getTwoRegionDomesticUseTransactions}
\alias{getTwoRegionDomesticUseTransactions}
\title{Load two-region domestic use transactions (intermediate consumption)
in state of interest (SoI) and its corresponding rest-of-US (RoUS) for a given year.}
\usage{
getTwoRegionDomesticUseTransactions(state, year, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, currently can only be "Summary",
theoretically can be "Detail", or "Sector" in future versions.}
}
\value{
A data.frame of two-region domestic use transactions.
}
\description{
Load two-region domestic use transactions (intermediate consumption)
in SoI and its corresponding RoUS for a given year.
}
