% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{getTwoRegionValueAdded}
\alias{getTwoRegionValueAdded}
\title{Load two-region value added in state of interest (SoI)
and its corresponding rest-of-US (RoUS) for a given year.}
\usage{
getTwoRegionValueAdded(state, year, iolevel)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}
}
\value{
A data.frame of two-region value added.
}
\description{
Load two-region value added of a SoI and its corresponding RoUS
for a given year.
}
