% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{loadBEAStateDatatoBEASummaryMapping}
\alias{loadBEAStateDatatoBEASummaryMapping}
\title{Load BEA State data (GVA and Employment) to BEA Summary mapping table}
\usage{
loadBEAStateDatatoBEASummaryMapping(dataname)
}
\arguments{
\item{dataname}{A string specifying name of the BEA state data}
}
\value{
The mapping table
}
\description{
Load BEA State data (GVA and Employment) to BEA Summary mapping table
}
