% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{loadStateIODataFile}
\alias{loadStateIODataFile}
\title{Load state IO data file from local data directory.}
\usage{
loadStateIODataFile(filename, ver = NULL)
}
\arguments{
\item{filename}{A string specifying filename, e.g. "State_Summary_Use_2017".}

\item{ver}{A string specifying version of the data, default is NULL, can be "v0.1.0".}
}
\value{
The pathname to the state IO data file.
}
\description{
Load state IO data file from local data directory.
}
