% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationFunctions.R
\name{validateTwoRegionLagainstOutput}
\alias{validateTwoRegionLagainstOutput}
\title{Validate Leontief matrix (L) of two-region model and final demand against
SoI and RoUS output.}
\usage{
validateTwoRegionLagainstOutput(state, year, iolevel, specs)
}
\arguments{
\item{state}{A text value specifying state of interest.}

\item{year}{A numeric value specifying the year of interest.}

\item{iolevel}{BEA sector level of detail, can be "Detail", "Summary", or "Sector".}

\item{specs}{A list of model specs including 'BaseIOSchema'}
}
\value{
A list of validation components and result.
}
\description{
Validate Leontief matrix (L) of two-region model and final demand against
SoI and RoUS output.
}
